Clustered database based on 90% identity
Clustering was done with mmseq using wsl (ubuntu). I created a python script that creates this clustered database file for you by automatically using the necessary commands (with subprocess library) in a WSL (Windows subsystem for Linux). So in order to run the script, you need to install a WSL (eg ubuntu) and install mmseqs in ubuntu.