New database with pept2taxa
By filtering the pept2taxa dataframe to only retain taxa's that match with at least 800 peptides in my de novo filtered and cleaned peptide sequences and by only using taxa that were defined on species level, I was able to obtain a database of around 500 MB (which is more or less optimal). I used an API call to query the proteomes of the taxa and downloaded the proteins in those proteomes by using their obtained proteome ID's and compiled all the proteins in a database. So, this may be a prototype worth testing in the lab.